refactor: 예외 폴더 분리
