+++ src/main/java/com/example/demo/base/exception/handler/OrderException.java
@@ -1,4 +1,4 @@
-package com.example.demo.base.exception;
+package com.example.demo.base.exception.handler;
 
 public class OrderException extends RuntimeException {
     public OrderException() {
