+++ src/main/java/com/example/demo/boundedContext/member/service/MemberService.java
@@ -1,7 +1,7 @@
 package com.example.demo.boundedContext.member.service;
 
 import com.example.demo.base.Role;
-import com.example.demo.base.exception.DataNotFoundException;
+import com.example.demo.base.exception.handler.DataNotFoundException;
 import com.example.demo.base.security.social.inter.DivideOAuth2User;
 import com.example.demo.boundedContext.member.entity.Address;
 import com.example.demo.boundedContext.member.entity.Member;
@@ -12,7 +12,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
