+++ src/main/java/com/example/demo/boundedContext/order/entity/OrderDetail.java
@@ -1,8 +1,7 @@
 package com.example.demo.boundedContext.order.entity;
 
 import com.example.demo.base.entity.BaseEntity;
-import com.example.demo.base.exception.NotEnoughProductCount;
-import com.example.demo.base.exception.OrderException;
+import com.example.demo.base.exception.handler.NotEnoughProductCount;
 import com.example.demo.boundedContext.product.entity.Product;
 import com.fasterxml.jackson.annotation.JsonBackReference;
