+++ src/test/java/com/example/demo/boundedContext/member/controller/MemberControllerTest.java
@@ -1,6 +1,6 @@
 package com.example.demo.boundedContext.member.controller;
 
-import com.example.demo.base.exception.DataNotFoundException;
+import com.example.demo.base.exception.handler.DataNotFoundException;
 import com.example.demo.boundedContext.member.entity.Member;
 import com.example.demo.boundedContext.member.service.MemberService;
 import org.junit.jupiter.api.DisplayName;
