+++ src/test/java/com/example/demo/boundedContext/order/service/OrderServiceTest.java
@@ -1,6 +1,6 @@
 package com.example.demo.boundedContext.order.service;
 
-import com.example.demo.base.exception.OrderException;
+import com.example.demo.base.exception.handler.OrderException;
 import com.example.demo.boundedContext.member.entity.Member;
 import com.example.demo.boundedContext.member.repository.MemberRepository;
 import com.example.demo.boundedContext.order.dto.OrderRequest;
